requirements.txt
streamlit
pandas
numpy
websocket-client
statsmodels
plotly
